services/realtime-gateway/requirements.txt
# Realtime Gateway Dependencies

# FastAPI and ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0

# WebSocket & Socket.IO
python-socketio==5.10.0
websockets==12.0

# Redis
redis==5.0.1
aioredis==2.0.1

# Authentication
python-jose[cryptography]==3.3.0
PyJWT==2.8.0

# Data Validation & Serialization
pydantic==2.5.0
pydantic-settings==2.1.0

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2

# Logging & Monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Testing (development)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1